Removed repetitive code xScale, yScale modules, modularized LineChart, updated types
