+++ src/App.tsx
@@ -96,7 +96,7 @@ function App() {
         />
 
         <LineChart
-          height="300"
+          height="100%"
           data={unemployment}
           xKey="date"
           xDataType="date"
@@ -110,7 +110,7 @@ function App() {
           yAxisLabel="Unemployment"
         />
         <LineChart
-          height="300"
+          height="500"
           data={portfolio}
           xKey="date"
           xDataType="date"
