+++ src/charts/BarChart/BarChartBody.tsx
@@ -4,7 +4,7 @@ import * as d3 from "d3";
 import ContinuousAxis from "../../components/ContinuousAxis";
 import DiscreteAxis from "../../components/DiscreteAxis";
 import { transformSkinnyToWide } from "../../utils";
-import { BarProps, ColorScale, Data } from "../../../types";
+import { BarChartProps, ColorScale, Data } from "../../../types";
 import {
   getXAxisCoordinates,
   getYAxisCoordinates,
@@ -30,7 +30,7 @@ const BarChartBody = ({
   xAxisLabel,
   yAxisLabel,
   colorScheme = d3.schemeCategory10,
-}: BarProps<number>): JSX.Element => {
+}: BarChartProps<number>): JSX.Element => {
   const margin = useMemo(
     () => getMargins(xAxis, yAxis, xAxisLabel, yAxisLabel),
     [xAxis, yAxis, xAxisLabel, yAxisLabel]
